added Toolbar, changed logic in Categories Page
