+++ src/components/app-toolbar/AppToolbar.styles.ts
@@ -1,0 +1,12 @@
+import { commonShadow } from '~/styles/app-theme/custom-shadows'
+
+export const styles = {
+  container: {
+    backgroundColor: 'basic.white', 
+    display: 'flex',
+    alignItems: 'center',
+    boxShadow: commonShadow,
+    mb: '50px',
+    p: '30px 45px'
+  }
+}
