+++ src/hooks/use-show-more.tsx
@@ -1,12 +1,14 @@
 import { Dispatch, SetStateAction, useMemo } from 'react'
 
-const useShowMore = <T,>(
+interface UseShowMoreProps<T> {
   limit: number,
   increaseCount: number,
   setLimit: Dispatch<SetStateAction<number>>,
   loading: boolean,
   response: { data: T[] }
-) => {
+}
+
+const useShowMore = <T,>({ limit, increaseCount, setLimit, loading, response }: UseShowMoreProps<T>) => {
   const isExpandable = useMemo(() => !loading && limit <= response.data.length, [limit, loading, response])
 
   const showMore = () => {
