--- src/pages/categories/Categories.styles.js
@@ -1,6 +1,0 @@
-export const styles = {
-  container: {
-    pt: 6,
-    flex: 1
-  }
-}
