+++ tests/unit/hooks/use-show-more.spec.js
@@ -1,26 +1,40 @@
 import { renderHook, act } from '@testing-library/react-hooks'
 
 import useShowMore from '~/hooks/use-show-more'
-import { mockCategories } from '~tests/unit/hooks/mock-categories'
+
+const setLimitMock = vi.fn()
+
+const props = {
+  limit: 1,
+  increaseCount: 2,
+  setLimit: setLimitMock,
+  loading: false,
+  response: {
+    data: [
+      { _id: '1', name: 'test' },
+      { _id: '2', name: 'test2' }
+    ]
+  }
+}
 
 describe('useShowMore custom hook', () => {
-  it('should do expand items by value from step', () => {
-    const { result } = renderHook(() => useShowMore(mockCategories, 2, 2))
+  it('should call setLimit', () => {
+    const { result } = renderHook(() => useShowMore({ ...props }))
 
     act(() => {
       result.current.showMore()
     })
 
-    expect(result.current.items.length).toBe(4)
+    expect(setLimitMock).toHaveBeenCalled()
   })
 
-  it('should assign expandable value to false, if there is no more items to show', () => {
-    const { result } = renderHook(() => useShowMore(mockCategories, 3, 2))
+  it('should return call show more and change isExpandable to true', () => {
+    const { result } = renderHook(() => useShowMore({ ...props }))
 
     act(() => {
       result.current.showMore()
     })
 
-    expect(result.current.expandable).toBeFalsy()
+    expect(result.current.isExpandable).toBeTruthy()
   })
 })
